requirements.txt
apache-airflow==2.10.5
beautifulsoup4==4.12.3
bokeh==3.3.4
boto3==1.34.156
DateTime==5.4
docker==7.0.0
factor-analyzer==0.5.1
fuzzywuzzy==0.18.0
google-api-python-client==2.123.0
httplib2==0.22.0
ipykernel==6.29.3
json-normalize==1.0.1
jupyter==1.0.0
kubernetes==31.0.0
lifelines==0.29.0
markdown==3.6
matplotlib==3.8.4
oauth2client==4.1.3
openpyxl==3.1.2
pandas==2.2.2
pendulum==3.0.0
pip==24.0
plotly==5.19.0
pmdarima==2.0.4
pre-commit==3.6.2
pyarrow==15.0.0
python-dotenv==1.0.1
requests==2.31.0
ruptures==1.1.9
scikit-learn==1.6.1
scikit-survival==0.24.0
scipy==1.12.0
seaborn==0.13.2
snowflake>=1.0.0
snowflake-connector-python[pandas,secure-local-storage]==3.14.0
snowflake-sqlalchemy==1.5.1
statsmodels==0.14.1
tabpy==2.11.0
xgboost==3.0.0
yellowbrick==1.5
yfinance==0.2.54